Mgc remove

+++ plugins/org.projectforge.plugins.datatransfer/src/main/kotlin/org/projectforge/plugins/datatransfer/rest/DataTransferPersonalBox.kt
@@ -30,10 +30,10 @@ import org.projectforge.plugins.datatransfer.DataTransferAreaDO
 import org.projectforge.rest.dto.AttachmentsSupport
 import org.projectforge.rest.dto.BaseDTO
 import org.projectforge.rest.dto.User
-import javax.persistence.Transient
+import jakarta.persistence.Transient
 
 class DataTransferPersonalBox(
-  id: Int? = null,
+  id: Long? = null,
   var user: User? = null,
   var internalLink: String? = null,
   override var attachmentsCounter: Int? = null,
