+++ projectforge-wicket/src/main/java/org/projectforge/renderer/custom/Formatter.java
@@ -29,8 +29,8 @@ import org.projectforge.business.timesheet.TimesheetDO;
 import org.projectforge.business.timesheet.TimesheetFilter;
 import org.projectforge.framework.persistence.user.api.ThreadLocalUserContext;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.List;
 import java.util.Map;
 
@@ -41,14 +41,14 @@ public abstract class Formatter
 {
   /**
    * Takes a Map of Parameters and returns a Map of data for the PDF Renderer
-   * 
+   *
    * @param parameters
    * @return
    */
-  public abstract Map<String, Object> getData(List<TimesheetDO> timeSheets, Integer taskId, HttpServletRequest request,
+  public abstract Map<String, Object> getData(List<TimesheetDO> timeSheets, Long taskId, HttpServletRequest request,
       HttpServletResponse response, TimesheetFilter actionFilter);
 
-  public Map<String, Object> getData(final List<TimesheetDO> timeSheets, final Integer taskId, final Request request, final Response response,
+  public Map<String, Object> getData(final List<TimesheetDO> timeSheets, final Long taskId, final Request request, final Response response,
       final TimesheetFilter actionFilter)
       {
     return getData(timeSheets, taskId, (HttpServletRequest) request.getContainerRequest(),
